apps/api/pyproject.toml
[project]
name = "tempo-api"
version = "0.1.0"
description = "Tempo personal audio server"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.1.0",
  "alembic>=1.13.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "python-multipart>=0.0.9",
  "orjson>=3.10.0",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt>=2.9.0",
  "email-validator>=2.1.0",
  "bcrypt>=4.0.1,<4.1.0",
]

[project.optional-dependencies]
media = ["mutagen>=1.47.0"]

test = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
